# docs/swagger.yaml
basePath: /
definitions:
...
  model.Session:
    properties:
      created_at:
        type: string
      device_id:
        type: string
      expires_at:
        type: string
      id:
        type: string
      ip_address:
        type: string
      is_current_session:
        type: boolean
      revoked_at:
        type: string
      tenant_id:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
...
  model.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      phone:
        type: string
      role:
        type: string
      tenant_id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.Error:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.Meta:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      request_id:
        type: string
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.Error'
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
...
  service.DispatcherLoginInput:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  service.DriverLoginInput:
    properties:
      phone:
        type: string
      pin:
        type: string
    required:
    - phone
    - pin
    type: object
  service.LoginOutput:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/model.User'
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@logistics.et
    name: API Support
  description: API for managing logistics operations including shipments, drivers,
    vehicles, and tracking
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ethiopian Logistics Tracking Platform API
  version: "1.0"
paths:
  /auth/login/dispatcher:
    post:
      consumes:
      - application/json
      description: Authenticate a dispatcher or admin using email and password
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.DispatcherLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Dispatcher login
      tags:
      - auth
  /auth/login/driver:
    post:
      consumes:
      - application/json
      description: Authenticate a driver using phone number and PIN
      parameters:
      - description: Login credentials
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/service.DriverLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      summary: Driver login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout and revoke the current session
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Get a new access token using a refresh token
      parameters:
      - description: Refresh token
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.LoginOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Refresh access token
      tags:
      - auth
...
  /sessions:
    get:
      description: Get all active sessions for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Session'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - sessions
  /sessions/{id}:
    delete:
      description: Revoke a specific session by ID
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Revoke a session
      tags:
      - sessions
  /sessions/others:
    delete:
      consumes:
      - application/json
      description: Revoke all sessions except the current one
      parameters:
      - description: Current session ID
        in: body
        name: input
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Revoke other sessions
      tags:
      - sessions
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"